Fix handling of last curve.
Logic bug in how split_painter_path handles last curve caused it to not be included. This was observable when using plotting order modes other than normal or reversed.
Blade offset had same bug, but it's harder to trigger because at stage of processing path usually ends with linear move.